requirements.txt
flask
gunicorn
PyMuPDF
python-docx
scikit-learn
pandas
google-cloud-firestore  # New: For storing usage data
sendgrid                # New: For sending emails (replace if using a different provider)
